refactor(transfer): split receiver/tests.rs by surface area
Decomposes the 4049-line receiver/tests.rs into a tests/ directory
keyed by transfer surface. Shared fixtures move into tests/support.rs;
each surface (file_list, delta_apply, hard_links, symlinks_and_devices,
partial_resume, errors_and_timeouts) lives in its own file with a
focused module doc.

Two files (file_list.rs, errors_and_timeouts.rs) still exceed the
650-line cap and carry TODO comments pointing at the next round of
splits. Test content is unchanged - only module structure moves.
